requirements.txt
google-genai==1.15.0
streamlit==1.45.1